# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "dssd-unlp-2021-grupo11-monitoreo-service",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "description": "Login con username y password",
                "operationId": "authLogin",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succesful login",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "401 Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Unauthorized"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Logout",
                "description": "Logout",
                "operationId": "authLogout",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/activeCase": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Get active cases",
                "description": "Get active cases",
                "operationId": "getActiveCases",
                "responses": {
                    "200": {
                        "description": "Active Cases",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "401 Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Unauthorized"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                }
            }
        },
        "/api/activeCaseCount": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Get active cases count",
                "description": "Get active cases count",
                "operationId": "getActiveCaseCount",
                "responses": {
                    "200": {
                        "description": "Active Cases Count",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "401 Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Unauthorized"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                }
            }
        },
        "/api/archivedCase": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Get archived cases",
                "description": "Get archived cases",
                "operationId": "getArchivedCases",
                "responses": {
                    "200": {
                        "description": "Archived Cases",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "401 Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Unauthorized"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                }
            }
        },
        "/api/archivedCaseCount": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Get archived cases count",
                "description": "Get archived cases count",
                "operationId": "getArchivedCaseCount",
                "responses": {
                    "200": {
                        "description": "Archived Cases Count",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "401 Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Unauthorized"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                }
            }
        },
        "/api/averageCaseTime": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Get average resolution time",
                "description": "Get average resolution time",
                "operationId": "getAverageCaseTime",
                "responses": {
                    "200": {
                        "description": "Average case time",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "401 Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Unauthorized"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                }
            }
        },
        "/api/process": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Get processes",
                "description": "Get processes",
                "operationId": "getProcesses",
                "responses": {
                    "200": {
                        "description": "Processes",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "401 Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Unauthorized"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                }
            }
        },
        "/api/user": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Get users",
                "description": "Get users",
                "operationId": "getUsers",
                "responses": {
                    "200": {
                        "description": "Users",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "401 Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Unauthorized"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Token Based",
                "name": "Token Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}